Move file cache bypass stuff into a mixin.
And also use it in file-provider adapter
[skip ci]
