+++ addon/adapters/file-provider.js
@@ -1,6 +1,7 @@
 import OsfAdapter from './osf-adapter';
+import FileCacheBypassMixin from 'ember-osf/mixins/file-cache-bypass';
 
-export default OsfAdapter.extend({
+export default OsfAdapter.extend(FileCacheBypassMixin, {
     pathForType() {
         return 'files';
     }
